chore: Update Dockerfile to use Node.js version 20.9.0
This commit updates the Dockerfile to use Node.js version 20.9.0 instead of the previous version 21-bullseye-slim. This change ensures compatibility with the required dependencies for the sharp library used for image processing. Additionally, the command for running the application in development mode has been removed, as it is not needed for exposing port 4000. This simplifies the Dockerfile and improves clarity.